test(core): split css.test.ts into flat-convention files
Two describe blocks → two files. The sb-prefix file keeps its
beforeAll with a comment noting the perf exception (~1s loadProject
would otherwise fire per-test). The empty-prefix test inlines its setup.
Shared helpers lifted to test/_helpers.ts.
